+++ apps/astarte_pairing/lib/astarte_pairing_web/metrics.ex
@@ -1,7 +1,7 @@
 #
 # This file is part of Astarte.
 #
-# Copyright 2017-2018 Ispirata Srl
+# Copyright 2020 Ispirata Srl
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,25 +16,8 @@
 # limitations under the License.
 #
 
-defmodule Astarte.Pairing do
-  @moduledoc false
-
-  use Application
-
-  alias Astarte.Pairing.Config
-  alias Astarte.Pairing.RPC.Handler
-
-  alias Astarte.RPC.Protocol.Pairing, as: Protocol
-
-  def start(_type, _args) do
-    Config.init!()
-
-    children = [
-      {Astarte.RPC.AMQP.Server, [amqp_queue: Protocol.amqp_queue(), handler: Handler]},
-      Astarte.PairingWeb.Metrics.Supervisor,
-      {Astarte.Pairing.CredentialsSecret.Cache, []}
-    ]
-
-    Supervisor.start_link(children, strategy: :one_for_one)
+defmodule Astarte.PairingWeb.Metrics do
+  def setup do
+    Astarte.PairingWeb.Metrics.PrometheusExporter.setup()
   end
 end