+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.11.0-beta.2] - Unreleased
 ### Added
+- [pairing] Add Prometheus instrumenters and exporters.
 - [pairing_api] Add Prometheus instrumenters and exporters
 - [realm_management_api] Add Prometheus instrumenters and exporters
 - [housekeeping_api] Add Prometheus instrumenters and exporters